requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
transformers==4.50.0
pillow==10.3.0
python-multipart==0.0.19
aiofiles==23.2.1
torch
pytesseract==0.3.10
sqlalchemy~=2.0.40
pydantic~=2.10.6
databases
python-dotenv
pyodbc
asyncpg

passlib~=1.7.4